% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-add-assay.R
\name{add_assay,OncoExperiment,OncoAssay-method}
\alias{add_assay,OncoExperiment,OncoAssay-method}
\title{Add an OncoAssay to an OncoExperiment}
\usage{
\S4method{add_assay}{OncoExperiment,OncoAssay}(object, assay, name = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{object}{An `OncoExperiment` object.}

\item{assay}{An object inheriting from `OncoAssay`.}

\item{name}{Optional character string specifying the assay name. If `NULL`,
`assay@name` is used.}

\item{overwrite}{Logical value indicating whether to overwrite an existing
assay with the same name.}

\item{...}{Additional arguments. Currently unused.}
}
\value{
The updated `OncoExperiment` object.
}
\description{
Adds an object inheriting from `OncoAssay` to the `assays` slot of an
`OncoExperiment`.
}
