% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-add-assay.R
\name{add_assay,OncoExperiment,SummarizedExperiment-method}
\alias{add_assay,OncoExperiment,SummarizedExperiment-method}
\title{Add an assay experiment to an OncoExperiment}
\usage{
\S4method{add_assay}{OncoExperiment,SummarizedExperiment}(object, assay, name = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{object}{An `OncoExperiment` object.}

\item{assay}{An assay object, such as a `SummarizedExperiment`,
`ExpressionAssay`, or `SingleCellExperiment`.}

\item{name}{A single character string specifying the experiment name.}

\item{overwrite}{Logical value indicating whether to overwrite an existing
experiment with the same name.}

\item{...}{Additional arguments. Currently unused.}
}
\value{
The updated `OncoExperiment` object.
}
\description{
Adds a named assay experiment to the `experiments()` of an `OncoExperiment`.
}
