% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-add-assay.R
\name{add_assay}
\alias{add_assay}
\title{Add an assay to an object}
\usage{
add_assay(object, assay, name = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{object}{The object to which the assay will be added.}

\item{assay}{The assay object to add.}

\item{name}{Optional character string specifying the assay name. If `NULL`,
the assay object's existing `name` slot is used.}

\item{overwrite}{Logical value indicating whether to overwrite an existing
assay with the same name.}

\item{...}{Additional arguments passed to methods.}
}
\value{
The updated object.
}
\description{
Adds an assay object to a container object.
}
