% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-OncoExperiment.R
\name{assay,OncoExperiment-method}
\alias{assay,OncoExperiment-method}
\title{Extract an assay from an OncoExperiment}
\usage{
\S4method{assay}{OncoExperiment}(object, name, ...)
}
\arguments{
\item{object}{An `OncoExperiment` object.}

\item{name}{A single character string specifying the name of the assay to
extract.}

\item{...}{Additional arguments. Currently unused.}
}
\value{
An object inheriting from `OncoAssay`.
}
\description{
Extracts a single assay from the `assays` slot of an `OncoExperiment` by
name.
}
\examples{
\dontrun{
expr <- assay(object, "Expression")
}

}
