% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-OncoExperiment.R
\name{assay}
\alias{assay}
\title{Extract an assay}
\usage{
assay(object, name, ...)
}
\arguments{
\item{object}{An object containing assays.}

\item{name}{A single character string specifying the assay name.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An assay object.
}
\description{
Extracts a single assay from an object by name.
}
