% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-OncoExperiment.R
\name{assays,OncoExperiment-method}
\alias{assays,OncoExperiment-method}
\title{List assay names in an OncoExperiment}
\usage{
\S4method{assays}{OncoExperiment}(object, ...)
}
\arguments{
\item{object}{An `OncoExperiment` object.}

\item{...}{Additional arguments. Currently unused.}
}
\value{
A character vector of assay names.
}
\description{
Returns the names of assays stored in the `assays` slot of an
`OncoExperiment`.
}
\examples{
\dontrun{
assays(object)
}

}
