% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-OncoExperiment.R
\name{assays}
\alias{assays}
\title{List assay names}
\usage{
assays(object, ...)
}
\arguments{
\item{object}{An object containing assays.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character vector of assay names.
}
\description{
Returns the names of assays stored in an object.
}
