% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-donut.R
\name{.apply_onco_plot_theme}
\alias{.apply_onco_plot_theme}
\title{Acquire the plotting theme from the environment}
\usage{
.apply_onco_plot_theme(base_size = 12)
}
\arguments{
\item{base_size}{Base font size.}
}
\value{
A ggplot theme.
}
\description{
If `theme_onco()` is not available, returns a generic ggplot minimal theme.
}
\keyword{internal}
