% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-metadata.R
\name{.find_sidecar_file}
\alias{.find_sidecar_file}
\title{Find a metadata sidecar file beside an assay file}
\usage{
.find_sidecar_file(path, filename)
}
\arguments{
\item{path}{Path to the primary assay file.}

\item{filename}{Metadata filename to find.}
}
\value{
A normalized character path if found, otherwise `NULL`.
}
\description{
Find a metadata sidecar file beside an assay file
}
\keyword{internal}
