% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-assays.R
\name{.get_loadable_assay_registry}
\alias{.get_loadable_assay_registry}
\title{Get loadable assay registry}
\usage{
.get_loadable_assay_registry(assay_type = NULL)
}
\arguments{
\item{assay_type}{Optional character vector of assay types to keep.}
}
\value{
A tibble containing loadable assay registry rows.
}
\description{
Returns a subset of the assay registry containing only rows that can be loaded
into assay objects.
}
\keyword{internal}
