% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-load-assays.R
\name{.resolve_files_to_load}
\alias{.resolve_files_to_load}
\title{Returns a tibble subset of the loadable assay registry based on the specified assay type.}
\usage{
.resolve_files_to_load(path, assay_type = NULL)
}
\arguments{
\item{path}{A character string specifying the directory or specific file path where the assay data is located. The function will look for files in this location that match the expected naming conventions for the specified assay type.}

\item{assay_type}{A character string specifying the type of assay being loaded (e.g., "Expression", "PRISM", etc.). This will determine the specific OncoAssay used and correct loading pattern. If NULL, all assay types in the registry will be considered.}
}
\value{
A tibble containing the loadable assay registry entries that match the specified assay type.
}
\description{
Returns a tibble subset of the loadable assay registry based on the specified assay type.
}
\keyword{internal}
