% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-donut.R
\name{.resolve_plot_donut_assay}
\alias{.resolve_plot_donut_assay}
\title{Resolve assay argument for plot_donut}
\usage{
.resolve_plot_donut_assay(assay)
}
\arguments{
\item{assay}{Assay name or enum value.}
}
\value{
A single character string.
}
\description{
Resolve assay argument for plot_donut
}
\keyword{internal}
