% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-compare-drug-sensitivities.R
\docType{class}
\name{DrugSensitivityComparison-class}
\alias{DrugSensitivityComparison-class}
\title{Drug sensitivity comparison result}
\description{
Drug sensitivity comparison result
}
\keyword{internal}
