% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-compare-drug-sensitivites.R
\name{export}
\alias{export}
\title{Export comparison results}
\usage{
export(object, ...)
}
\arguments{
\item{object}{A `DrugSensitivityComparison` object.}

\item{...}{Additional arguments passed to the export method.}
}
\description{
Export comparison results
}
\keyword{internal}
