% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-OncoExperiment-load_assays.R
\name{fetch_available_assays}
\alias{fetch_available_assays}
\title{Fetch available assays from DepMap API
This function interacts with the DepMap API to retrieve a list of available assays for a
given release. It returns a character vector of assay names that can be used with the `load_depmap_assay` method.}
\usage{
fetch_available_assays()
}
\value{
A character vector of available assay names.
}
\description{
Fetch available assays from DepMap API
This function interacts with the DepMap API to retrieve a list of available assays for a
given release. It returns a character vector of assay names that can be used with the `load_depmap_assay` method.
}
