% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-assays.R
\name{get_supported_assays}
\alias{get_supported_assays}
\title{List all supported assays}
\usage{
get_supported_assays()
}
\value{
A tribble
}
\description{
Returns all supported assays in a tribble to be used as a lookup table.
}
