% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-assays.R
\name{get_supported_types}
\alias{get_supported_types}
\title{List supported assay types}
\usage{
get_supported_types()
}
\value{
A character vector of supported assay types.
}
\description{
Returns all supported assay types declared in the assay registry.
}
