% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ExpressionAssay.R
\name{internal-ExpressionAssay-validity}
\alias{internal-ExpressionAssay-validity}
\title{Validity method for ExpressionAssay}
\description{
Ensures that the object contains a numeric assay named `"expression"` with
row and column names.
}
\keyword{internal}
