% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ProteinExpressionAssay.R
\name{internal-ProteinExpressionAssay-validity}
\alias{internal-ProteinExpressionAssay-validity}
\title{Validity method for ProteinExpressionAssay}
\description{
Ensures that the object contains a numeric assay named
`"protein_expression"` with row and column names.
}
\keyword{internal}
