% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-load-assays.R
\name{load_assays,OncoExperiment-method}
\alias{load_assays,OncoExperiment-method}
\title{Load assay data into an OncoExperiment object}
\usage{
\S4method{load_assays}{OncoExperiment}(
  object,
  assay_type = NULL,
  data = NULL,
  metadata = NULL,
  identifier = NULL,
  name = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{object}{An `OncoExperiment` object to which the assay will be added.}

\item{assay_type}{A character string of the type of assay being loaded. Use the `AssayTypes` enum to easily select a valid assay type.}

\item{data}{A character string specifying the path to the data file. This is required.}

\item{metadata}{A character string specifying the path to the metadata file. This is optional but highly recommended.}

\item{identifier}{A character string of the column name where the unique identifier is found that links your data to the metadata. If not supplied, the first column is used.}

\item{name}{A character string of a unique name for this assay. If not supplied and an assay of the same type exists, they will be named sequentially.}

\item{overwrite}{A logical value indicating whether to overwrite an existing
assay with the same name.}
}
\value{
The updated `OncoExperiment` object.
}
\description{
This function loads and parses the specified data and metadata files and appends to the @assays
slot of the `OncoExperiment` object. You can view available assay types by printing `AssayTypes`.
}
