% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-load-assays.R
\name{load_assays}
\alias{load_assays}
\title{Load assay data into an OncoExperiment object
The function will coerce the assay into the corresponding OncoAssay class
before being added to the OncoExperiment object.}
\usage{
load_assays(object, assay_type = NULL, path = NULL, ...)
}
\arguments{
\item{object}{An OncoExperiment object to which the assay will be added.}

\item{assay_type}{A character string specifying the type of assay being loaded (e.g., "Expression", "PRISM", etc.). This will determine the specific OncoAssay used and correct loading pattern.}

\item{path}{A character string specifying the directory or specific file path where the assay data is located. The function will look for files in this location that match the expected naming conventions for the specified assay type.}
}
\value{
An updated OncoExperiment with loaded assays in the @assays slot.
}
\description{
Load assay data into an OncoExperiment object
The function will coerce the assay into the corresponding OncoAssay class
before being added to the OncoExperiment object.
}
