% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-OncoExperiment-load_depmap_assay.R
\name{load_depmap_assay}
\alias{load_depmap_assay}
\title{downloadAssays method for OncoExperiment objects
Interacts with the DepMap API to download assay data.}
\usage{
\S4method{load_depmap_assay}{OncoExperiment}(object, release, assay = "Expression", ...)
}
\description{
downloadAssays method for OncoExperiment objects
Interacts with the DepMap API to download assay data.
}
