% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-ModelMetadataFields.R
\name{ModelMetadataFields}
\alias{ModelMetadataFields}
\title{ModelMetadataFields}
\usage{
ModelMetadataFields
}
\description{
An enumeration of the metadata fields available in an ExpressionAssay.
}
