% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OncoAssay.R
\docType{class}
\name{OncoAssay}
\alias{OncoAssay}
\title{OncoAssay virtual S4 class definition
This class serves as a base class for different types of assays in oncology experiments,
such as gene expression matrices, mutation data, or other experimental data. It is designed
to hold the assay data, metadata, and any additional information relevant to the assay.}
\value{
An object of class OncoAssay
}
\description{
OncoAssay virtual S4 class definition
This class serves as a base class for different types of assays in oncology experiments,
such as gene expression matrices, mutation data, or other experimental data. It is designed
to hold the assay data, metadata, and any additional information relevant to the assay.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{A list of metadata information related to the assay, such as sample annotations, experimental conditions, or other relevant information.}

\item{\code{data}}{A matrix containing the assay data, such as gene expression values or mutation counts.}

\item{\code{row_data}}{A list of metadata information for the rows of the data matrix, such as gene annotations or mutation information.}

\item{\code{col_data}}{A list of metadata information for the columns of the data matrix, such as sample annotations or experimental conditions.}

\item{\code{layers}}{A list of additional data layers related to the assay, such as normalized data, log-transformed data, or other derived data matrices.}

\item{\code{name}}{A character string representing the name of the assay.}
}}

