% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-donut.R
\name{plot_donut,OncoExperiment-method}
\alias{plot_donut,OncoExperiment-method}
\title{Create a donut plot from model metadata}
\usage{
\S4method{plot_donut}{OncoExperiment}(
  object,
  variable,
  assay,
  variable_expr = substitute(variable),
  show_na = FALSE,
  ...
)
}
\arguments{
\item{object}{An `OncoExperiment` object.}

\item{variable}{Metadata field to plot.}

\item{assay}{Assay name to use.}

\item{variable_expr}{Captured unevaluated variable expression. Internal.}

\item{show_na}{Logical. Whether missing values should be shown as `"Missing"`.}

\item{...}{Additional arguments passed to `.apply_onco_plot_theme()`.}
}
\value{
A `ggplot` object.
}
\description{
Generates a donut plot that shows the distribution of categorical metadata
fields from a selected assay in an `OncoExperiment`.
}
