% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-donut.R
\name{plot_donut}
\alias{plot_donut}
\title{Create a donut plot}
\usage{
plot_donut(object, variable, assay, variable_expr = substitute(variable), ...)
}
\arguments{
\item{object}{An `OncoExperiment` object.}

\item{variable}{Metadata field to plot. Can be unquoted, a character string,
or an enum value such as `ModelMetadataFields$OncotreeLineage`.}

\item{assay}{A single character string specifying which assay experiment to
use. For example, `"Expression"` or `"PRISM"`.}

\item{variable_expr}{Captured unevaluated variable expression. Internal.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A `ggplot` object.
}
\description{
Creates a donut plot from assay-level metadata stored in an `OncoExperiment`.
}
\examples{
\dontrun{
plot_donut(exp, ModelMetadataFields$OncotreeLineage, assay = "Expression")
plot_donut(exp, "Sex", assay = "Expression")
plot_donut(exp, OncotreeLineage, assay = "Expression")
}

}
