% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-compare-drug-sensitivities.R
\name{SensitivityUnits}
\alias{SensitivityUnits}
\title{Drug sensitivity response units
LFC = log fold change, 
AUC = area under the curve,
IC50 = half maximal inhibitory concentration}
\usage{
SensitivityUnits
}
\description{
Drug sensitivity response units
LFC = log fold change, 
AUC = area under the curve,
IC50 = half maximal inhibitory concentration
}
