% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ExpressionAssay.R
\name{show,ExpressionAssay-method}
\alias{show,ExpressionAssay-method}
\title{Show an ExpressionAssay object}
\usage{
\S4method{show}{ExpressionAssay}(object)
}
\arguments{
\item{object}{An `ExpressionAssay` object.}
}
\value{
Invisibly returns `NULL`.
}
\description{
Show an ExpressionAssay object
}
