% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OncoExperiment.R
\name{show,OncoExperiment-method}
\alias{show,OncoExperiment-method}
\title{Show an OncoExperiment object}
\usage{
\S4method{show}{OncoExperiment}(object)
}
\arguments{
\item{object}{An `OncoExperiment` object.}
}
\value{
Invisibly returns `NULL`.
}
\description{
Show an OncoExperiment object
}
