% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ProteinExpressionAssay.R
\name{show,ProteinExpressionAssay-method}
\alias{show,ProteinExpressionAssay-method}
\title{Show an ExpressionAssay object}
\usage{
\S4method{show}{ProteinExpressionAssay}(object)
}
\arguments{
\item{object}{An `ExpressionAssay` object.}
}
\value{
Invisibly returns `NULL`.
}
\description{
Show an ExpressionAssay object
}
