% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-report.R
\name{summarize_agent_results}
\alias{summarize_agent_results}
\title{Summarize enrichment agent results}
\usage{
summarize_agent_results(result_table, agent_name, result_type, top_n = 10)
}
\arguments{
\item{result_table}{Enrichment result table.}

\item{agent_name}{Name of the agent.}

\item{result_type}{Type of result being summarized.}

\item{top_n}{Number of top terms to include.}
}
\value{
Character summary.
}
\description{
Summarize enrichment agent results
}
