% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-assays.R
\name{supported_assays}
\alias{supported_assays}
\title{Mapping of supported assay types, their corresponding assay classes, filename patterns, and metadata dependencies.}
\usage{
supported_assays
}
\description{
These are the only files downloaded from DepMap, inferred, and loadable in
their assay-specific classes.
}
\keyword{internal}
