% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-theme.R
\name{theme_onco}
\alias{theme_onco}
\title{OncoProfilingTools ggplot theme}
\usage{
theme_onco(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme.
}
\description{
A minimal theme for publication-style oncology profiling figures.
}
