% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-assays.R
\name{AssayTypes}
\alias{AssayTypes}
\title{An enumeration of supported assay types}
\usage{
AssayTypes
}
\description{
Helpful for intelligent code completion and avoiding typos in assay types.
}
\examples{
AssayTypes # you can see all supported assays available
AssayTypes$RNA # it's just a string at the end of the day


}
