% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-assays.R
\name{AssayTypes}
\alias{AssayTypes}
\title{An enumeration of suppported assay types.
Helpful for intelligent code completion and to avoid typos in assay type strings.}
\usage{
AssayTypes
}
\description{
Curious what types are available? Just print `AssayTypes`.
}
\examples{
AssayTypes$Expression
AssayTypes$ProteinExpression
}
