% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-load-assays.R
\name{.load_assay}
\alias{.load_assay}
\title{Load a single assay using a registered loader}
\usage{
.load_assay(path, filename, assay_type, OncoAssay_class, loader, ...)
}
\arguments{
\item{path}{Character scalar. Path to the assay file.}

\item{filename}{Character scalar. File name being loaded.}

\item{assay_type}{Character scalar. Project-level assay type.}

\item{OncoAssay_class}{Character scalar. Expected returned assay class.}

\item{loader}{Character scalar. Registered loader function name.}

\item{...}{Additional arguments passed to the registered loader.}
}
\value{
An assay object of the expected class.
}
\description{
Load a single assay using a registered loader
}
\keyword{internal}
