% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-download-assays.R
\name{download_assays}
\alias{download_assays}
\title{Interacts with the DepMap API to download the specified supported assays.
It will download the assay files for the latest release of that type in store them in the `.cache` directory.}
\usage{
\S4method{download_assays}{OncoExperiment}(object, assay_type, ...)
}
\arguments{
\item{object}{An OncoExperiment object to store the downloaded assay data.}

\item{assay_type}{The type of assay to download, e.g. "expression". See supported assay types with `list_supported_assays()`.}
}
\value{
An updated OncoExperiment object with the downloaded assay data stored in the @assays slot.
}
\description{
If the files exist already, it will skip their download.
}
