% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ExpressionAssay.R,
%   R/class-ProteinExpressionAssay.R
\docType{class}
\name{ExpressionAssay}
\alias{ExpressionAssay}
\title{ExpressionAssay S4 class definition
This class is designed to hold gene expression assay data, including the expression matrix, metadata, and additional information relevant to the assay.
It inherits from the OncoAssay virtual class. Usually, you won't create an ExpressionAssay manually but use a helper function for your data source.}
\value{
An object of class ExpressionAssay

An object of class ExpressionAssay
}
\description{
Typical use case is for DepMap gene expression data.

Typical use case is for DepMap gene expression data.
}
\section{Slots}{

\describe{
\item{\code{unit}}{A character string representing the unit of measurement for the expression data, such as "TPM", "RPKM", or "raw counts".}

\item{\code{normalized}}{A logical value indicating whether the expression data has been normalized (TRUE) or is in raw form (FALSE).}

\item{\code{unit}}{A character string representing the unit of measurement for the expression data, such as "TPM", "RPKM", or "raw counts".}

\item{\code{normalized}}{A logical value indicating whether the expression data has been normalized (TRUE) or is in raw form (FALSE).}
}}

