% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-download-assays.R
\name{fetch_all_assay_urls}
\alias{fetch_all_assay_urls}
\title{Entry URL's from the API are not valid download links.
Instead, we can use them to get a csv table listing all
files available for download to find the correct download URL}
\usage{
fetch_all_assay_urls(object = NULL)
}
\arguments{
\item{object}{An optional OncoExperiment object to store the downloaded assay data. Default is NULL.}
}
\value{
A data frame of available assay names.
}
\description{
You can provide an OncoExperiment object to store the data.
Caching is useful as it takes a few secodnds to download the data.
This is full optional.
}
\keyword{internal}
