% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-OncoExperiment-load_assays.R
\name{find_assay_url}
\alias{find_assay_url}
\title{Find the download URL for a specific assay file based on the release and filename}
\usage{
find_assay_url(release, filename)
}
\arguments{
\item{release}{The DepMap release, e.g. "25Q3"}

\item{filename}{The name of the file to download, e.g. "depmap_25Q3_expression.csv"}
}
\value{
The download URL for the specified file
}
\description{
Find the download URL for a specific assay file based on the release and filename
}
\keyword{internal}
