% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-assays.R
\name{get_supported_types}
\alias{get_supported_types}
\title{A helper function to list all supported assay types that OncoExperiment can load.}
\usage{
get_supported_types()
}
\description{
A helper function to list all supported assay types that OncoExperiment can load.
}
