% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-load-assays.R
\name{load_assays,OncoExperiment-method}
\alias{load_assays,OncoExperiment-method}
\title{Load assay data into an OncoExperiment object}
\usage{
\S4method{load_assays}{OncoExperiment}(object, assay_type = NULL, path = ".cache", overwrite = FALSE, ...)
}
\arguments{
\item{object}{An OncoExperiment object to which the assay will be added.}

\item{assay_type}{A character string specifying the type of assay being loaded (e.g., "Expression", "PRISM", etc.). This will determine the specific OncoAssay used and correct loading pattern.}

\item{path}{A character string specifying the directory or specific file path where the assay data is located. The function will look for files in this location that match the expected naming conventions for the specified assay type.}

\item{overwrite}{A logical value indicating whether to overwrite an existing assay of the same type in the OncoExperiment object. Default is FALSE, which will prevent overwriting and throw an error if an assay of the same type already exists.}
}
\description{
The function will coerce the assay into the corresponding OncoAssay class
before being added to the OncoExperiment object. The assay type specified will determine the specific OncoAssay used and the correct loading pattern.
}
