% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-load-assays.R
\name{load_assays,OncoExperiment-method}
\alias{load_assays,OncoExperiment-method}
\title{Load assay data into an OncoExperiment object}
\usage{
\S4method{load_assays}{OncoExperiment}(object, assay_type = NULL, path = ".cache", overwrite = FALSE, ...)
}
\arguments{
\item{object}{An `OncoExperiment` object to which the assay will be added.}

\item{assay_type}{A character vector specifying the assay type(s) being loaded,
such as `"Expression"` or `"PRISM"`.}

\item{path}{A character string specifying the directory or specific file path
where the assay data is located.}

\item{overwrite}{A logical value indicating whether to overwrite an existing
assay with the same name.}

\item{...}{Additional arguments passed to assay-specific loader functions.}
}
\value{
The updated `OncoExperiment` object.
}
\description{
Loads one or more supported assay files into an `OncoExperiment`.
}
\details{
The assay type determines which registered loader is used. Each loader returns
an assay object, such as an `ExpressionAssay`, which is then added to the
`OncoExperiment`.
}
