% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-load-assays.R
\name{load_assays}
\alias{load_assays}
\title{Load assay data into an OncoExperiment object
The function will coerce the assay into the corresponding OncoAssay class
before being added to the OncoExperiment object.}
\usage{
load_assays(object, assay_type = NULL, data = NULL, metadata = NULL, ...)
}
\arguments{
\item{object}{An OncoExperiment object to which the assay will be added.}

\item{assay_type}{A character string of the type of assay being loaded. Use the `AssayTypes` enum to easily select a valid assay type.}

\item{data}{A character string specifying specific file path where the assay data is located.}

\item{metadata}{A character string specifying the path to the metadata file. This is optional but highly recommended.}
}
\value{
An updated OncoExperiment with loaded assays in the @assays slot.
}
\description{
Load assay data into an OncoExperiment object
The function will coerce the assay into the corresponding OncoAssay class
before being added to the OncoExperiment object.
}
