% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OncoExperiment.R
\docType{class}
\name{OncoExperiment-class}
\alias{OncoExperiment-class}
\title{OncoExperiment S4 class definition
This class is designed to hold all relevant data and results for an oncology experiment,
including assays, models, graphs, and metadata.}
\value{
An object of class OncoExperiment
}
\description{
OncoExperiment S4 class definition
This class is designed to hold all relevant data and results for an oncology experiment,
including assays, models, graphs, and metadata.
}
\section{Slots}{

\describe{
\item{\code{Optional}}{project_name A character string representing the name of the project.}

\item{\code{version}}{A character string representing the version of the package.}

\item{\code{assays}}{A list of OncoAssay objects, such as gene expression matrices or other experimental data.}

\item{\code{models}}{A list of OncoModel objects, which may include fitted models, predictions, or other analytical results.}

\item{\code{graphs}}{A list of OncoGraph objects, which may include ggplot2 objects or other visualizations related to the experiment.}

\item{\code{metadata}}{A list of metadata information, such as sample annotations, experimental conditions, or other relevant information.}
}}

\examples{
# Create an empty OncoExperiment object
experiment <- OncoExperiment(project_name = "My Oncology Experiment")
}
