% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OncoExperiment.R
\name{OncoExperiment}
\alias{OncoExperiment}
\title{Create an empty OncoExperiment object}
\usage{
OncoExperiment(project_name = "OncoExperiment Project")
}
\arguments{
\item{project_name}{A character string representing the project name.}
}
\value{
An empty object of class `OncoExperiment`.
}
\description{
Creates an empty oncology-focused `MultiAssayExperiment` subclass.
}
\details{
Assays and metadata should be added after construction using package helper
functions such as `load_assays()` and `add_assay()`.
}
\examples{
experiment <- OncoExperiment()

}
