% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OncoExperiment.R
\name{show,OncoExperiment-method}
\alias{show,OncoExperiment-method}
\title{Show method for OncoExperiment objects.
This method defines how OncoExperiment objects are displayed when printed.}
\usage{
\S4method{show}{OncoExperiment}(object)
}
\description{
Show method for OncoExperiment objects.
This method defines how OncoExperiment objects are displayed when printed.
}
