% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-subset.R
\name{subset,OncoExperiment-method}
\alias{subset,OncoExperiment-method}
\title{Subset an OncoExperiment}
\usage{
\S4method{subset}{OncoExperiment}(object, subset, select = TRUE, ...)
}
\arguments{
\item{object}{An `OncoExperiment` object to be subset.}

\item{subset}{A logical expression indicating which rows to keep.}

\item{select}{A logical expression indicating which columns to keep.}
}
\value{
A new `OncoExperiment` object containing the subset of the original data.
}
\description{
Allow the user the subset an OncoExperiment by a given condition,
similar to the subset function in Seurat. The condition will be evaluated
and the heavy lifitng of subsetting the `MultiAssayExperiment` within an `OncoExperiment`
will be takne care of.
}
