% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-OncoExperiment-load_assays.R
\name{supported_assays}
\alias{supported_assays}
\title{Mapping of supported assay types, their corresponding OncoAssay classes, and filename patterns
These are the only files downloaded from DepMap, inferred, and loadable in their assay-specific classes.}
\usage{
supported_assays
}
\description{
Mapping of supported assay types, their corresponding OncoAssay classes, and filename patterns
These are the only files downloaded from DepMap, inferred, and loadable in their assay-specific classes.
}
\keyword{internal}
