% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-assays.R
\docType{data}
\name{supported_assays}
\alias{supported_assays}
\title{Mapping of supported assay types, their corresponding OncoAssay classes, and filename patterns
These are the only files downloaded from DepMap, inferred, and loadable in their assay-specific classes.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 5 columns.
}
\usage{
supported_assays
}
\description{
Mapping of supported assay types, their corresponding OncoAssay classes, and filename patterns
These are the only files downloaded from DepMap, inferred, and loadable in their assay-specific classes.
}
\keyword{internal}
